requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fonttools==4.54.1
h11==0.14.0
hdbscan==0.8.39
httpcore==1.0.6
httpx==0.27.2
idna==3.10
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-core==0.3.10
langchain-ollama==0.2.0
langsmith==0.1.132
llvmlite==0.43.0
lxml==5.3.0
matplotlib==3.9.2
numba==0.60.0
numpy==2.0.2
ollama==0.3.3
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
